Reorganize directory and update config
