+++ pyproject.toml
@@ -1,16 +1,26 @@
 [build-system]
-requires = ["scikit-build-core", "pybind11"]
+requires = ["scikit-build-core[pyproject]", "pybind11"]
 build-backend = "scikit_build_core.build"
 
 [project]
 # See https://setuptools.pypa.io/en/latest/userguide/quickstart.html for more project configuration options.
 name = "bang-gpu"
 readme = "README.md"
-version = "0.0.1"
+version = "0.0.1-alpha.1"
 
 requires-python = ">=3.8"
 dependencies = ["python-libsbml"]
 license = { file = "LICENSE" }
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+]
+
+__all__ = ["bang-gpu"]
 
 [project.urls]
 Homepage = "https://github.com/pavozayac/bang"
@@ -18,27 +28,31 @@ Repository = "https://github.com/pavozayac/bang"
 Changelog = "https://github.com/pavozayac/bang/blob/main/CHANGELOG.md"
 # Documentation = "https://bang.readthedocs.io/"
 
+[tool.scikit-build]
+wheel.packages = ["src/bang"]
+cmake.verbose = true
+
 [project.optional-dependencies]
 dev = [
-  "pybind11",
-  "ruff",
-  "mypy>=1.0,<1.5",
-  "black>=23.0,<24.0",
-  "isort>=5.12,<5.13",
-  "pytest",
-  "pytest-sphinx",
-  "pytest-cov",
-  "twine>=1.11.0",
-  "build",
-  "setuptools",
-  "wheel",
-  "Sphinx>=4.3.0,<7.1.0",
-  "furo==2023.7.26",
-  "myst-parser>=1.0,<2.1",
-  "sphinx-copybutton==0.5.2",
-  "sphinx-autobuild==2021.3.14",
-  "sphinx-autodoc-typehints==1.23.3",
-  "packaging",
+    "pybind11",
+    "ruff",
+    "mypy>=1.0,<1.5",
+    "black>=23.0,<24.0",
+    "isort>=5.12,<5.13",
+    "pytest",
+    "pytest-sphinx",
+    "pytest-cov",
+    "twine>=1.11.0",
+    "build",
+    "setuptools",
+    "wheel",
+    "Sphinx>=4.3.0,<7.1.0",
+    "furo==2023.7.26",
+    "myst-parser>=1.0,<2.1",
+    "sphinx-copybutton==0.5.2",
+    "sphinx-autobuild==2021.3.14",
+    "sphinx-autodoc-typehints==1.23.3",
+    "packaging",
 ]
 
 [tool.setuptools.packages.find]
