+++ src/bang/core/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+# from ._gpu_stable import german_gpu_run, initialise_PBN
+from .PBN import PBN, load_from_file
+
+__all__ = ["PBN", "load_from_file"]
