+++ src/bang/core/boolFunc.py
