+++ src/bang/core/parseSBML.py
