+++ src/bang/ffi/GPU_stable.cu
