+++ src/bang/ffi/GermanGPU.h
