+++ src/bang/ffi/shared/German.cc
