+++ src/bang/ffi/shared/German.h
