+++ src/bang/ffi/shared/ltqnorm.h
