+++ src/bang/version.py
